Fix a bug on deleteFile function on FileController, the function did not eliminate the file register from the database when deleting a file

+++ app/Http/Controllers/FileController.php
@@ -195,6 +195,15 @@ class FileController extends Controller
             $errorMessage = "You don't own that";
             return redirect()->back()->with('errorMessage',$errorMessage);
         }
+
+        // delete from data base
+        $file = File::where("id","=",$id,"and","user_id","=",Auth::user()->id)->first();
+        if ($file == null) {
+            $errorMessage = "File is corrupted";
+            return redirect()->back()->with('errorMessage',$errorMessage);
+        }
+        // delete from database
+        $file->delete();
         $query = Storage::disk('public')->delete(substr($file[0]->path,8));
         if ($query === true) {
             $file[0]->delete();
